pyproject.toml
[project]
name = "neurolight-prototype"
version = "1.0.0"
description = "A powerful PySide6 desktop application for processing and analyzing large TIF image stacks with scientific rigor"
requires-python = ">=3.9"
dependencies = [
    "PySide6>=6.6.0",
    "opencv-python>=4.8.0",
    "Pillow>=10.0.0",
    "ultralytics>=8.0.0",
    "numpy>=1.24.0",
    "scipy>=1.11.0",
    "matplotlib>=3.8.0",
    "statsmodels>=0.14.0",
    "imageio>=2.31.0",
    "pystackreg>=0.2.8",
    "tifffile>=2023.1.0",
    "scikit-image>=0.21.0",
]

[project.scripts]
neurolight = "src.main:main"

#  Pyrefly configuration
[tool.pyrefly]
project-includes = ["**/*.py", "**/*.pyi"]
project-excludes = ["**/node_modules/**", "**/__pycache__/**", "**/.venv/**"]
use-ignore-files = true
python-version = "3.9"
search-path = ["."]

[project.optional-dependencies]
test = ["pytest>=7.0.0", "pytest-cov>=4.0.0"]
# Note: dev dependencies are primarily managed via [dependency-groups] dev
# This is kept for backward compatibility with tools that use optional-dependencies
dev = ["ruff>=0.14.3", "mypy>=1.0.0", "pyrefly"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = ["ruff>=0.14.3", "mypy>=1.0.0", "pyrefly"]

[tool.hatch.build.targets.wheel]
packages = ["src"]

# Ruff configuration
[tool.ruff]
line-length = 100
target-version = "py39"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "N",   # pep8-naming
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "SIM", # flake8-simplify
    "ARG", # flake8-unused-arguments
    "RUF", # Ruff-specific rules
]
ignore = [
    "E501",   # Line too long (handled by formatter)
    "B008",   # Do not perform function calls in argument defaults
    "ARG001", # Unused function argument
    "ARG002", # Unused method argument
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["ARG", "S101"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "auto"

[tool.ruff.lint.isort]
known-first-party = ["src"]
split-on-trailing-comma = true
